requirements.txt

torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.66.0
pyyaml>=6.0
